requirements.txt
ipython
jupyterlab
tqdm
ipywidgets
chromadb-client
ollama
networkx